+++ src/main.rs
@@ -74,6 +74,7 @@ cfg_if::cfg_if! {
         mod top_items_summary;
         mod date;
         mod clean_unref;
+        mod verify;
 
         // use
         use crate::cache::caches::{Cache, RegistrySuperCache};
@@ -88,6 +89,7 @@ cfg_if::cfg_if! {
         use crate::top_items_summary::*;
         use crate::clean_unref::*;
         use crate::cli::{CargoCacheCommands};
+        //use crate::verify;
     }
 }
 
@@ -465,6 +467,36 @@ fn main() {
         CargoCacheCommands::OnlyDryRun => {
             if !size_changed {
                 eprintln!("Warning: there is nothing to be dry run!");
+            }
+        }
+        CargoCacheCommands::Verify {
+            clean_corrupted,
+            dry_run,
+        } => {
+            println!("Verifying cache, this may take some time...\n");
+            if let Err(failed_verifications) = verify::verify_crates(&mut registry_sources_caches) {
+                eprintln!("\n");
+                failed_verifications
+                    .iter()
+                    .for_each(|diff| println!("{}", diff.details()));
+                eprintln!(
+                    "\nFound {} possible corrupted sources.",
+                    failed_verifications.len()
+                );
+
+                if clean_corrupted {
+                    verify::clean_corrupted(
+                        &mut registry_sources_caches,
+                        &failed_verifications,
+                        dry_run,
+                    );
+                } else {
+                    println!("Hint: use `cargo cache verify --clean-corrupted` to remove them.");
+                }
+
+                std::process::exit(1)
+            } else {
+                std::process::exit(0);
             }
         }
         _ => (),